# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "description": "Public API error categories (see ``api/errors.py`` for the envelope).",
        "enum": [
          "VALIDATION",
          "AUTHENTICATION",
          "AUTHORIZATION",
          "SCOPE_DENIED",
          "NOT_FOUND",
          "CONFLICT",
          "RATE_LIMIT",
          "DEPENDENCY_UNAVAILABLE",
          "TIMEOUT",
          "RETRYABLE_PROCESSING",
          "CORRUPT_SOURCE",
          "INSUFFICIENT_EVIDENCE",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorEnvelope": {
        "$defs": {
          "ErrorBody": {
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "Stable machine-readable category."
              },
              "details": {
                "additionalProperties": true,
                "title": "Details",
                "type": "object"
              },
              "message": {
                "description": "Human-readable summary. Never contains source text.",
                "title": "Message",
                "type": "string"
              },
              "retryable": {
                "description": "Whether the same request may succeed if retried.",
                "title": "Retryable",
                "type": "boolean"
              },
              "trace_id": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "default": null,
                "description": "Correlates with logs/traces.",
                "title": "Trace Id"
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "title": "ErrorBody",
            "type": "object"
          },
          "ErrorCode": {
            "description": "Public API error categories (see ``api/errors.py`` for the envelope).",
            "enum": [
              "VALIDATION",
              "AUTHENTICATION",
              "AUTHORIZATION",
              "SCOPE_DENIED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMIT",
              "DEPENDENCY_UNAVAILABLE",
              "TIMEOUT",
              "RETRYABLE_PROCESSING",
              "CORRUPT_SOURCE",
              "INSUFFICIENT_EVIDENCE",
              "INTERNAL"
            ],
            "title": "ErrorCode",
            "type": "string"
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "OutcomeRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "note": "user accepted the quote",
            "scope": {
              "agent_run_id": "run_01J8ZK",
              "thread_id": "thr_01J8ZK7Q9V3W2X1Y0ZABCDEFGH"
            },
            "success": true
          }
        ],
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeBody"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "OutcomeRequest",
        "type": "object"
      },
      "OutcomeResponse": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "run_id",
          "success",
          "source"
        ],
        "title": "OutcomeResponse",
        "type": "object"
      },
      "PlanRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "available_tools": [
              {
                "name": "pricing.lookup_price"
              },
              {
                "name": "crm.update_quote"
              }
            ],
            "scope": {
              "agent_run_id": "run_01J8ZK",
              "thread_id": "thr_01J8ZK7Q9V3W2X1Y0ZABCDEFGH"
            },
            "task": "update quote Q-1183 with EMEA price for SKU-22"
          }
        ],
        "properties": {
          "available_tools": {
            "items": {
              "$ref": "#/components/schemas/DeclaredTool"
            },
            "title": "Available Tools",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeBody"
          },
          "task": {
            "maxLength": 4000,
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "PlanRequest",
        "type": "object"
      },
...
      "ProceduresResponse": {
        "properties": {
          "procedures": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Procedures",
            "type": "array"
          }
        },
        "required": [
          "procedures"
        ],
        "title": "ProceduresResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "trusted_dev mode only",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "jwt / gcp_iam modes",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "Memory Service API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs/{run_id}/outcome": {
      "post": {
        "operationId": "set_run_outcome_v1_runs__run_id__outcome_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Client request id (generated when absent).",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Correlation id shared by related requests.",
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Trace id for OpenTelemetry correlation.",
            "in": "header",
            "name": "X-Trace-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Idempotency key for safe retries of persistent writes.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTHENTICATION",
                    "details": {},
                    "message": "Missing or invalid credentials",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCOPE_DENIED",
                    "details": {},
                    "message": "Access denied",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "details": {},
                    "message": "Request validation failed",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DEPENDENCY_UNAVAILABLE",
                    "details": {},
                    "message": "PostgreSQL unavailable",
                    "retryable": true,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "PostgreSQL unavailable"
          }
        },
        "summary": "Label an agent run successful or not (only successful runs validate a procedure)",
        "tags": [
          "tools"
        ]
      }
    },
...
    "/v1/tools/plan": {
      "post": {
        "operationId": "plan_tools_v1_tools_plan_post",
        "parameters": [
          {
            "description": "Client request id (generated when absent).",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Correlation id shared by related requests.",
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Trace id for OpenTelemetry correlation.",
            "in": "header",
            "name": "X-Trace-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Idempotency key for safe retries of persistent writes.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTHENTICATION",
                    "details": {},
                    "message": "Missing or invalid credentials",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCOPE_DENIED",
                    "details": {},
                    "message": "Access denied",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "details": {},
                    "message": "Request validation failed",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DEPENDENCY_UNAVAILABLE",
                    "details": {},
                    "message": "PostgreSQL unavailable",
                    "retryable": true,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "PostgreSQL unavailable"
          }
        },
        "summary": "The best-known validated chain for a task, as an ordered plan with bindings",
        "tags": [
          "tools"
        ]
      }
    },
    "/v1/tools/procedures": {
      "get": {
        "operationId": "list_procedures_v1_tools_procedures_get",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "default": "",
              "title": "Task",
              "type": "string"
            }
          },
          {
            "description": "The agent whose procedures to list. Required to see agent-scoped invocations: they are recorded against principal 'agent:<id>', and lineage cannot travel in a header the way tenant/workspace/user do.",
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "default": "",
              "description": "The agent whose procedures to list. Required to see agent-scoped invocations: they are recorded against principal 'agent:<id>', and lineage cannot travel in a header the way tenant/workspace/user do.",
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "description": "Narrow to one workspace.",
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "default": "",
              "description": "Narrow to one workspace.",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "description": "Client request id (generated when absent).",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Correlation id shared by related requests.",
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Trace id for OpenTelemetry correlation.",
            "in": "header",
            "name": "X-Trace-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProceduresResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTHENTICATION",
                    "details": {},
                    "message": "Missing or invalid credentials",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCOPE_DENIED",
                    "details": {},
                    "message": "Access denied",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "details": {},
                    "message": "Request validation failed",
                    "retryable": false,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DEPENDENCY_UNAVAILABLE",
                    "details": {},
                    "message": "PostgreSQL unavailable",
                    "retryable": true,
                    "trace_id": "req_01J8ZZZZZZZZZZZZZZZZZZZZZZ"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "PostgreSQL unavailable"
          }
        },
        "summary": "Procedures mined for a task pattern",
        "tags": [
          "tools"
        ]
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "current host",
      "url": "/"
    }
  ],
...
}